Update to E3 plus update formating

+++ src/main/java/io/nem/sdk/model/transaction/SecretProofTransaction.java
@@ -27,187 +27,255 @@ import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.util.Optional;
 
-/**
- * Secret proof transaction.
- */
+/** Secret proof transaction. */
 public class SecretProofTransaction extends Transaction {
-	private final HashType hashType;
-	private final String secret;
-	private final String proof;
-	private final Address recipient;
+  private final HashType hashType;
+  private final String secret;
+  private final String proof;
+  private final Address recipient;
 
-	/**
-	 * Constructor.
-	 *
-	 * @param networkType     Network type.
-	 * @param version         Transaction version.
-	 * @param deadline        Deadline to include the transaction.
-	 * @param maxFee          Max fee the sender will pay.
-	 * @param hashType        Hash algorithm secret is generated with.
-	 * @param recipient       Address of recipient.
-	 * @param secret          Seed proof hashed.
-	 * @param proof           Seed proof.
-	 * @param signature       Transaction Signature.
-	 * @param signer          Signer of the transaction.
-	 * @param transactionInfo Transaction info.
-	 */
-	public SecretProofTransaction(final NetworkType networkType, final Integer version, final Deadline deadline, final BigInteger maxFee,
-								  final HashType hashType, final Address recipient, final String secret, final String proof,
-								  final String signature, final PublicAccount signer, final TransactionInfo transactionInfo) {
-		this(networkType, version, deadline, maxFee, hashType, recipient, secret, proof, Optional.of(signature), Optional.of(signer),
-				Optional.of(transactionInfo));
-	}
+  /**
+   * Constructor.
+   *
+   * @param networkType Network type.
+   * @param version Transaction version.
+   * @param deadline Deadline to include the transaction.
+   * @param maxFee Max fee the sender will pay.
+   * @param hashType Hash algorithm secret is generated with.
+   * @param recipient Address of recipient.
+   * @param secret Seed proof hashed.
+   * @param proof Seed proof.
+   * @param signature Transaction Signature.
+   * @param signer Signer of the transaction.
+   * @param transactionInfo Transaction info.
+   */
+  public SecretProofTransaction(
+      final NetworkType networkType,
+      final Short version,
+      final Deadline deadline,
+      final BigInteger maxFee,
+      final HashType hashType,
+      final Address recipient,
+      final String secret,
+      final String proof,
+      final String signature,
+      final PublicAccount signer,
+      final TransactionInfo transactionInfo) {
+    this(
+        networkType,
+        version,
+        deadline,
+        maxFee,
+        hashType,
+        recipient,
+        secret,
+        proof,
+        Optional.of(signature),
+        Optional.of(signer),
+        Optional.of(transactionInfo));
+  }
 
-	/**
-	 * Constructor.
-	 *
-	 * @param networkType Network type.
-	 * @param version     Transaction version.
-	 * @param deadline    Deadline to include the transaction.
-	 * @param maxFee      Max fee the sender will pay.
-	 * @param hashType    Hash algorithm secret is generated with.
-	 * @param recipient   Address of recipient.
-	 * @param secret      Seed proof hashed.
-	 * @param proof       Seed proof.
-	 */
-	public SecretProofTransaction(final NetworkType networkType, final Integer version, final Deadline deadline, final BigInteger maxFee,
-								  final HashType hashType, final Address recipient, final String secret, final String proof) {
-		this(networkType, version, deadline, maxFee, hashType, recipient, secret, proof, Optional.empty(), Optional.empty(),
-				Optional.empty());
-	}
+  /**
+   * Constructor.
+   *
+   * @param networkType Network type.
+   * @param version Transaction version.
+   * @param deadline Deadline to include the transaction.
+   * @param maxFee Max fee the sender will pay.
+   * @param hashType Hash algorithm secret is generated with.
+   * @param recipient Address of recipient.
+   * @param secret Seed proof hashed.
+   * @param proof Seed proof.
+   */
+  public SecretProofTransaction(
+      final NetworkType networkType,
+      final Short version,
+      final Deadline deadline,
+      final BigInteger maxFee,
+      final HashType hashType,
+      final Address recipient,
+      final String secret,
+      final String proof) {
+    this(
+        networkType,
+        version,
+        deadline,
+        maxFee,
+        hashType,
+        recipient,
+        secret,
+        proof,
+        Optional.empty(),
+        Optional.empty(),
+        Optional.empty());
+  }
 
-	/**
-	 * Constructor.
-	 *
-	 * @param networkType     Network type.
-	 * @param version         Transaction version.
-	 * @param deadline        Deadline to include the transaction.
-	 * @param maxFee          Max fee the sender will pay.
-	 * @param hashType        Hash algorithm secret is generated with.
-	 * @param recipient       Address of recipient.
-	 * @param secret          Seed proof hashed.
-	 * @param proof           Seed proof.
-	 * @param signature       Transaction Signature.
-	 * @param signer          Signer of the transaction.
-	 * @param transactionInfo Transaction info.
-	 */
-	public SecretProofTransaction(final NetworkType networkType, final Integer version, final Deadline deadline, final BigInteger maxFee,
-								  final HashType hashType, final Address recipient, final String secret, final String proof,
-								  final Optional<String> signature, final Optional<PublicAccount> signer,
-								  final Optional<TransactionInfo> transactionInfo) {
-		super(TransactionType.SECRET_PROOF, networkType, version, deadline, maxFee, signature, signer, transactionInfo);
-		Validate.notNull(secret, "Secret must not be null.");
-		Validate.notNull(proof, "Proof must not be null.");
-		Validate.notNull(recipient, "Recipient must not be null.");
-		if (!HashType.Validator(hashType, secret)) {
-			throw new IllegalArgumentException("HashType and Secret have incompatible length or not hexadecimal string");
-		}
-		this.hashType = hashType;
-		this.secret = secret;
-		this.proof = proof;
-		this.recipient = recipient;
-	}
+  /**
+   * Constructor.
+   *
+   * @param networkType Network type.
+   * @param version Transaction version.
+   * @param deadline Deadline to include the transaction.
+   * @param maxFee Max fee the sender will pay.
+   * @param hashType Hash algorithm secret is generated with.
+   * @param recipient Address of recipient.
+   * @param secret Seed proof hashed.
+   * @param proof Seed proof.
+   * @param signature Transaction Signature.
+   * @param signer Signer of the transaction.
+   * @param transactionInfo Transaction info.
+   */
+  public SecretProofTransaction(
+      final NetworkType networkType,
+      final Short version,
+      final Deadline deadline,
+      final BigInteger maxFee,
+      final HashType hashType,
+      final Address recipient,
+      final String secret,
+      final String proof,
+      final Optional<String> signature,
+      final Optional<PublicAccount> signer,
+      final Optional<TransactionInfo> transactionInfo) {
+    super(
+        TransactionType.SECRET_PROOF,
+        networkType,
+        version,
+        deadline,
+        maxFee,
+        signature,
+        signer,
+        transactionInfo);
+    Validate.notNull(secret, "Secret must not be null.");
+    Validate.notNull(proof, "Proof must not be null.");
+    Validate.notNull(recipient, "Recipient must not be null.");
+    if (!HashType.Validator(hashType, secret)) {
+      throw new IllegalArgumentException(
+          "HashType and Secret have incompatible length or not hexadecimal string");
+    }
+    this.hashType = hashType;
+    this.secret = secret;
+    this.proof = proof;
+    this.recipient = recipient;
+  }
 
-	/**
-	 * Create a secret proof transaction object.
-	 *
-	 * @param deadline    Deadline to include the transaction.
-	 * @param hashType    Hash algorithm secret is generated with.
-	 * @param recipient   Address of recipient.
-	 * @param maxFee      Max fee the sender will pay.
-	 * @param secret      Seed proof hashed.
-	 * @param proof       Seed proof.
-	 * @param networkType Network type.
-	 * @return a SecretLockTransaction instance
-	 */
-	public static SecretProofTransaction create(final Deadline deadline, final HashType hashType, final Address recipient,
-												final BigInteger maxFee, final String secret, final String proof,
-												final NetworkType networkType) {
-		return new SecretProofTransaction(networkType, TransactionVersion.SECRET_PROOF.getValue(), deadline, maxFee, hashType, recipient,
-				secret, proof);
-	}
+  /**
+   * Create a secret proof transaction object.
+   *
+   * @param deadline Deadline to include the transaction.
+   * @param hashType Hash algorithm secret is generated with.
+   * @param recipient Address of recipient.
+   * @param maxFee Max fee the sender will pay.
+   * @param secret Seed proof hashed.
+   * @param proof Seed proof.
+   * @param networkType Network type.
+   * @return a SecretLockTransaction instance
+   */
+  public static SecretProofTransaction create(
+      final Deadline deadline,
+      final HashType hashType,
+      final Address recipient,
+      final BigInteger maxFee,
+      final String secret,
+      final String proof,
+      final NetworkType networkType) {
+    return new SecretProofTransaction(
+        networkType,
+        TransactionVersion.SECRET_PROOF.getValue(),
+        deadline,
+        maxFee,
+        hashType,
+        recipient,
+        secret,
+        proof);
+  }
 
-	/**
-	 * Returns the hash algorithm secret is generated with.
-	 *
-	 * @return the hash algorithm secret is generated with.
-	 */
-	public HashType getHashType() {
-		return hashType;
-	}
+  /**
+   * Returns the hash algorithm secret is generated with.
+   *
+   * @return the hash algorithm secret is generated with.
+   */
+  public HashType getHashType() {
+    return hashType;
+  }
 
-	/**
-	 * Returns the proof hashed.
-	 *
-	 * @return the proof hashed.
-	 */
-	public String getSecret() {
-		return secret;
-	}
+  /**
+   * Returns the proof hashed.
+   *
+   * @return the proof hashed.
+   */
+  public String getSecret() {
+    return secret;
+  }
 
-	/**
-	 * Returns proof.
-	 *
-	 * @return proof.
-	 */
-	public String getProof() {
-		return proof;
-	}
+  /**
+   * Returns proof.
+   *
+   * @return proof.
+   */
+  public String getProof() {
+    return proof;
+  }
 
-	/**
-	 * Serialized the transaction.
-	 *
-	 * @return bytes of the transaction.
-	 */
-	@Override
-	byte[] generateBytes() {
-		// Add place holders to the signer and signature until actually signed
-		final ByteBuffer signerBuffer = ByteBuffer.allocate(32);
-		final ByteBuffer signatureBuffer = ByteBuffer.allocate(64);
+  /**
+   * Serialized the transaction.
+   *
+   * @return bytes of the transaction.
+   */
+  @Override
+  byte[] generateBytes() {
+    // Add place holders to the signer and signature until actually signed
+    final ByteBuffer signerBuffer = ByteBuffer.allocate(32);
+    final ByteBuffer signatureBuffer = ByteBuffer.allocate(64);
 
-		SecretProofTransactionBuilder txBuilder =
-				SecretProofTransactionBuilder.create(new SignatureDto(signatureBuffer),
-						new KeyDto(signerBuffer), getNetworkVersion(),
-						EntityTypeDto.SECRET_PROOF_TRANSACTION,
-						new AmountDto(getFee().longValue()), new TimestampDto(getDeadline().getInstant()),
-						LockHashAlgorithmDto.rawValueOf((byte) hashType.getValue()),
-						new Hash256Dto(getSecretBuffer()),
-						new UnresolvedAddressDto(this.recipient.getByteBuffer()),
-						getProofBuffer());
-		return txBuilder.serialize();
-	}
+    SecretProofTransactionBuilder txBuilder =
+        SecretProofTransactionBuilder.create(
+            new SignatureDto(signatureBuffer),
+            new KeyDto(signerBuffer),
+            getNetworkVersion(),
+            EntityTypeDto.SECRET_PROOF_TRANSACTION,
+            new AmountDto(getFee().longValue()),
+            new TimestampDto(getDeadline().getInstant()),
+            LockHashAlgorithmDto.rawValueOf((byte) hashType.getValue()),
+            new Hash256Dto(getSecretBuffer()),
+            new UnresolvedAddressDto(this.recipient.getByteBuffer()),
+            getProofBuffer());
+    return txBuilder.serialize();
+  }
 
-	/**
-	 * Gets the embedded tx bytes.
-	 *
-	 * @return Embedded tx bytes
-	 */
-	@Override
-	byte[] generateEmbeddedBytes() {
-		EmbeddedSecretProofTransactionBuilder txBuilder =
-				EmbeddedSecretProofTransactionBuilder.create(new KeyDto(getSignerBytes().get()), getNetworkVersion(),
-						EntityTypeDto.SECRET_PROOF_TRANSACTION,
-						LockHashAlgorithmDto.rawValueOf((byte) hashType.getValue()),
-						new Hash256Dto(getSecretBuffer()),
-						new UnresolvedAddressDto(this.recipient.getByteBuffer()),
-						getProofBuffer());
-		return txBuilder.serialize();
-	}
+  /**
+   * Gets the embedded tx bytes.
+   *
+   * @return Embedded tx bytes
+   */
+  @Override
+  byte[] generateEmbeddedBytes() {
+    EmbeddedSecretProofTransactionBuilder txBuilder =
+        EmbeddedSecretProofTransactionBuilder.create(
+            new KeyDto(getSignerBytes().get()),
+            getNetworkVersion(),
+            EntityTypeDto.SECRET_PROOF_TRANSACTION,
+            LockHashAlgorithmDto.rawValueOf((byte) hashType.getValue()),
+            new Hash256Dto(getSecretBuffer()),
+            new UnresolvedAddressDto(this.recipient.getByteBuffer()),
+            getProofBuffer());
+    return txBuilder.serialize();
+  }
 
-	private ByteBuffer getSecretBuffer() {
-		final ByteBuffer secretBuffer = ByteBuffer.allocate(32);
-		secretBuffer.put(Hex.decode(secret));
-		return secretBuffer;
-	}
+  private ByteBuffer getSecretBuffer() {
+    final ByteBuffer secretBuffer = ByteBuffer.allocate(32);
+    secretBuffer.put(Hex.decode(secret));
+    return secretBuffer;
+  }
 
-	/**
-	 * Gets proof buffer
-	 *
-	 * @return Proof buffer.
-	 */
-	private ByteBuffer getProofBuffer() {
-		final byte[] proofBytes = Hex.decode(proof);
-		final ByteBuffer proofBuffer = ByteBuffer.wrap(proofBytes);
-		return proofBuffer;
-	}
+  /**
+   * Gets proof buffer
+   *
+   * @return Proof buffer.
+   */
+  private ByteBuffer getProofBuffer() {
+    final byte[] proofBytes = Hex.decode(proof);
+    final ByteBuffer proofBuffer = ByteBuffer.wrap(proofBytes);
+    return proofBuffer;
+  }
 }
